+++ pilot/pkg/networking/core/v1alpha3/gateway.go
